% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00Classes.R
\docType{class}
\name{cles-class}
\alias{cles-class}
\title{Class "cles"}
\description{
Defines the object returned by different methods in the package
}
\section{Slots}{


\describe{
\item{\code{Call}:}{function call}
\item{\code{Statistics}:}{common language effect size values}
\item{\code{Significance}:{standard errors and t-test value}}
}
}

\section{Methods}{


\describe{
\item{summary}{\code{signature(object = "cles")}}
}
}

\author{
Timi Niemensivu \email{timinie@utu.fi}

Jari Metsämuuronen \email{jari.metsamuuronen@gmail.com}
}
\keyword{classes}
