% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLES-package.R
\name{CLES-package}
\alias{CLES-package}
\title{Common Language Effect Size}
\description{
Common Language Effect Sizes
}
\author{
Timi Niemensivu \email{timinie@utu.fi}

Jari Metsämuuronen \email{jari.metsamuuronen@gmail.com}
}
\keyword{package}
