% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_PS.R
\name{ordinal_PS}
\alias{ordinal_PS}
\alias{ordinal_A}
\title{Function to calculate ordinal A and ordinal PS effect size.
As ordinal A and ordinal PS are equal (see Metsämuuronen 2024), effect size
calculated here will be called ordinal A even if it is also ordinal PS.}
\usage{
ordinal_PS(
  x,
  y = NULL,
  conf.level = 0.05,
  error.type = "normal",
  alternative = "2-sided"
)
}
\arguments{
\item{x}{either a table or a vector of a same length as y}

\item{y}{if x is supplied as a vector, a vector same length as x}

\item{conf.level}{confidence level used for confidence interval, default is 0.05}

\item{error.type}{type of asymptotic standard error to be used in calculation
of the p-value. Default is "normal" with other option being "uniform" for
the uniform distribution based standard error and hypothesis testing}

\item{alternative}{type of alternative hypothesis used for calculating p-value.
Either "2-sided" (default) or "1-sided"}
}
\description{
Function to calculate ordinal A and ordinal PS effect size.
As ordinal A and ordinal PS are equal (see Metsämuuronen 2024), effect size
calculated here will be called ordinal A even if it is also ordinal PS.
}
\examples{

\dontrun{

data(PHD_data)
ordinal_A_res <- ordinal_A(PHD_data$g1, PHD_data$X)
summary(ordinal_A_res)
}
}
\references{
Metsämuuronen, Jari. (2024). Two new common language estimators of effect size: Somers' delta and Goodman-Kruskal gamma as bases for nonparametric effect sizes. (Preprint) 10.13140/RG.2.2.14774.31045.
}
\seealso{
\code{\link{PHD}} \code{\link{PHG}} \code{\link{ordinal_d}}
}
\author{
Timi Niemensivu \email{timinie@utu.fi}

Jari Metsämuuronen \email{jari.metsamuuronen@gmail.com}
}
\keyword{common}
\keyword{effect}
\keyword{language}
\keyword{sizes}
