% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00Classes.R
\docType{class}
\name{PHD-class}
\alias{PHD-class}
\title{Class "PHD"}
\description{
Defines the object returned from \code{\link{PHD}}.
}
\section{Slots}{


\describe{
\item{\code{Call}:}{function call }
\item{\code{Statistics}:}{D and PHD values}
\item{\code{Significance}:{standard errors and t-test value}}
}
}

\section{Methods}{


\describe{
\item{summary}{\code{signature(object = "PHD")}}
}
}

\author{
Timi Niemensivu \email{timinie@utu.fi}

Jari Metsämuuronen \email{jari.metsamuuronen@gmail.com}
}
\keyword{classes}
