% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHD.R
\name{PHD}
\alias{PHD}
\title{Function to calculate PHD effect size}
\usage{
PHD(
  x,
  y = NULL,
  conf.level = 0.05,
  error.type = "normal",
  alternative = "2-sided"
)
}
\arguments{
\item{x}{either a table or a vector of a same length as y}

\item{y}{if x is supplied as a vector, a vector same length as x}

\item{conf.level}{confidence level used for confidence interval, default is 0.05}

\item{error.type}{type of asymptotic standard error to be used in calculation
of the p-value. Default is "normal" with other option being "uniform" for
the uniform distribution based standard error and hypothesis testing}

\item{alternative}{type of alternative hypothesis used for calculating p-value.
Either "2-sided" (default) or "1-sided"}
}
\description{
Function to calculate PHD effect size
}
\examples{

\dontrun{

data(PHD_data)
PHD_res <- PHD(PHD_data$g1, PHD_data$X)
summary(PHD)
}
}
\seealso{
\code{\link{PHG}}
}
\author{
Timi Niemensivu \email{timinie@utu.fi}

Jari Metsämuuronen \email{jari.metsamuuronen@gmail.com}
}
\keyword{common}
\keyword{effect}
\keyword{language}
\keyword{sizes}
