% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEF-package.R
\docType{data}
\name{PHD_data}
\alias{PHD_data}
\title{Description of Science data}
\description{
A 4-item data set borrowed from \code{ltm} package in R, first example
of the \code{grm()} function. See more complete documentation therein, as
well as Karlheinz and Melich (1992).
}
\references{
Metsämuuronen, J. (2024). Two new nonparametric estimators of common language effect size.
\emph{Journal of Experimental Education} (In printing)
}
\author{
Jari Metsämuuronen \email{jari.metsamuuronen@gmail.com}
}
\keyword{data}
