% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLES-package.R
\docType{data}
\name{PHD_data}
\alias{PHD_data}
\title{Description of PHD example data}
\description{
Hypothetical dataset used in Metsämuuronen 2024 to demonstrate properties of PHD and PHG.
}
\examples{

\dontrun{

data(PHD_data)
PHD_res <- PHD(PHD_data$g1, PHD_data$X)
summary(PHD)
}
}
\references{
Metsämuuronen, Jari. (2024). Two new common language estimators of effect size: Somers' delta and Goodman-Kruskal gamma as bases for nonparametric effect sizes. (Preprint) 10.13140/RG.2.2.14774.31045.
}
\author{
Jari Metsämuuronen \email{jari.metsamuuronen@gmail.com}
}
\keyword{data}
